requirements.txt
Flask==3.1.0
Flask-wtf==1.2.2
Flask-SQLAlchemy==3.1.1
Flask_mail==0.9.1
Flask-Cors==5.0.1
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.4
requests==2.32.3
mysql-connector-python==9.2.0
cachetools==5.5.2
torch==2.6.0
setuptools==76.1.0
wheel==0.45.1
nltk==3.8.1